+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [1.2.0] - 2016-01-29
+## [1.3.0] - 2016-02-10
 ### Added
-- Repository manager to get a repository depending on your loaded modules and their repository manager configuration
+- Validator `ObjectExists` to determinate wetter a requires object/entity exists or not
 
 ### Changed 
 - Nothing
@@ -16,10 +16,17 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Nothing
 
 ### Fixed
-- Entity manager
+- Nothing
 
 ### Security
 - Nothing
+
+## [1.2.0] - 2016-01-29
+### Added
+- Repository manager to get a repository depending on your loaded modules and their repository manager configuration
+
+### Fixed
+- Entity manager
 
 ## [1.1.0] - 2016-01-26
 ### Added